Feature/improve design

+++ app/src/main/java/com/mir/panosdev/cookingrecipesmvp/mvp/presenter/RecipesPresenter.java
@@ -1,31 +1,25 @@
 package com.mir.panosdev.cookingrecipesmvp.mvp.presenter;
 
-import android.os.Handler;
 import android.util.Log;
 
 import com.mir.panosdev.cookingrecipesmvp.api.RecipesApiService;
-import com.mir.panosdev.cookingrecipesmvp.base.BasePresenter;
 import com.mir.panosdev.cookingrecipesmvp.mapper.RecipeMapper;
 import com.mir.panosdev.cookingrecipesmvp.mvp.model.recipes.Recipe;
 import com.mir.panosdev.cookingrecipesmvp.mvp.model.recipes.RecipesResponse;
 import com.mir.panosdev.cookingrecipesmvp.mvp.model.recipes.Storage;
-import com.mir.panosdev.cookingrecipesmvp.mvp.model.users.User;
-import com.mir.panosdev.cookingrecipesmvp.mvp.view.MainView;
-
+import com.mir.panosdev.cookingrecipesmvp.mvp.view.MainActivityMVP;
 import java.util.List;
-
 import javax.inject.Inject;
-
 import io.reactivex.Observable;
-import io.reactivex.Observer;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.annotations.NonNull;
+import io.reactivex.disposables.CompositeDisposable;
 import io.reactivex.disposables.Disposable;
+import io.reactivex.observers.DisposableObserver;
+import io.reactivex.schedulers.Schedulers;
 import retrofit2.Response;
 
-/**
- * Created by Panos on 3/18/2017.
- */
-
-public class RecipesPresenter extends BasePresenter<MainView> implements Observer<Response<RecipesResponse>> {
+public class RecipesPresenter implements MainActivityMVP.Presenter {
 
     @Inject
     protected RecipesApiService mRecipesApiService;
@@ -35,6 +29,8 @@ public class RecipesPresenter extends BasePresenter<MainView> implements Observe
     @Inject
     protected Storage mStorage;
 
+    private MainActivityMVP.MainView mainView;
+    private CompositeDisposable compositeDisposable = null;
 
     @Inject
     public RecipesPresenter() {
@@ -42,40 +38,55 @@ public class RecipesPresenter extends BasePresenter<MainView> implements Observe
 
     @Inject
     public void getRecipes() {
-//        getView().onShowDialog("Loading recipes....");
-        Observable<Response<RecipesResponse>> recipesResponseObservable = mRecipesApiService.getRecipes();
-        subscribe(recipesResponseObservable, this);
-        mStorage.dropDatabase();
+        if (mainView != null) {
+            mainView.onShowDialog("Loading recipes....");
+            mStorage.dropDatabase();
+            Observable<Response<RecipesResponse>> recipesResponseObservable = mRecipesApiService.getRecipes();
+            Disposable disposable = recipesResponseObservable.observeOn(AndroidSchedulers.mainThread())
+                    .subscribeOn(Schedulers.io())
+                    .subscribeWith(new DisposableObserver<Response<RecipesResponse>>() {
+                        @Override
+                        public void onNext(@NonNull Response<RecipesResponse> recipesResponseResponse) {
+                            List<Recipe> recipes = mRecipeMapper.mapRecipesAndStorage(mStorage, recipesResponseResponse.body().getRecipes());
+                            mainView.onClearItems();
+                            mainView.onRecipeLoaded(recipes);
+                        }
+
+                        @Override
+                        public void onError(@NonNull Throwable e) {
+                            mainView.onHideDialog();
+                            mainView.onShowToast("Error loading recipes " + e.getMessage());
+                            Log.e("ERROR_LOG", e.getMessage());
+                        }
+
+                        @Override
+                        public void onComplete() {
+                            mainView.onHideDialog();
+                        }
+
+                    });
+            if (compositeDisposable != null)
+                compositeDisposable.add(disposable);
+        }
     }
 
-    @Override
-    public void onSubscribe(Disposable d) {
-    }
-
-    @Override
-    public void onNext(Response<RecipesResponse> recipesResponse) {
-        List<Recipe> recipes = mRecipeMapper.mapRecipes(mStorage, recipesResponse.body().getRecipes());
-        getView().onClearItems();
-        getView().onRecipeLoaded(recipes);
-    }
 
     public void getRecipesFromDatabase() {
         List<Recipe> recipes = mStorage.getSavedRecipes();
-        getView().onClearItems();
-        getView().onRecipeLoaded(recipes);
-        getView().onNetworkUnavailableToast("Updating items from database...");
+        mainView.onClearItems();
+        mainView.onRecipeLoaded(recipes);
+        mainView.onNetworkUnavailableToast("Updating items from database...");
     }
 
     @Override
-    public void onError(Throwable e) {
-        getView().onHideDialog();
-        getView().onShowToast("Error loading recipes " + e.getMessage());
+    public void attachView(MainActivityMVP.MainView view) {
+        mainView = view;
     }
 
     @Override
-    public void onComplete() {
-        getView().onHideDialog();
-        getView().onShowToast("Sync completed!");
-
+    public void detachView() {
+        if (compositeDisposable != null) {
+            compositeDisposable.dispose();
+        }
     }
 }
